requirements.txt
tensorflow
numpy
pillow
scikit-learn
tk
